apps/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.25
openpyxl==3.1.2
python-multipart==0.0.9
